requirements.txt
streamlit
tensorflow
scikit-learn
joblib
pandas
numpy
Pillow
openpyxl